WebServices/WebService/requirements.txt
Django
django-crispy-forms==1.7.2
Pillow
python-magic
zeep==3.4.0
pathlib
argparse